+++ ocrd/decorators.py
@@ -57,7 +57,7 @@ def ocrd_cli_options(f):
         click.option('-w', '--working-dir', help="Working Directory"),
         click.option('-I', '--input-file-grp', help='File group(s) used as input.', default='INPUT'),
         click.option('-O', '--output-file-grp', help='File group(s) used as output.', default='OUTPUT'),
-        click.option('-g', '--group-id', help="mets:file GROUPID"),
+        click.option('-P', '--page-id', help="ID(s) of the pages to process"),
         click.option('-p', '--parameter', type=click.Path()),
         click.option('-J', '--dump-json', help="Dump tool description as JSON and exit", is_flag=True, default=False),
         loglevel_option,