+++ CHANGELOG.md
@@ -5,9 +5,11 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [0.14.0] - 2018-12-13
+
 Changed:
 
-  * :fire: Use mets:structMap[@TYPE="PHYSICAL"] instead of GROUPID to group by page, #221, OCR-D/spec#81
+  * :fire: Use mets:structMap[@TYPE="PHYSICAL"] instead of GROUPID to group by page, #221, #222, OCR-D/spec#81
   * :fire: Rename group_id -> page_id, throughout, OCR-D/spec#101
   * :fire: CLI: Rename --group-id -> --page-id (`-g`)